+++ emeraldcomponents/src/main/java/br/com/stone/emeraldcomponents/basic/label/LabelShapeHandler.kt
@@ -1,0 +1,5 @@
+package br.com.stone.emeraldcomponents.basic.label
+
+interface LabelShapeHandler {
+    fun setBackground(label: EmeraldLabel)
+}